remove duplicate relationshipCards view and use common one for snapshot and screening
