+++ app/javascript/containers/snapshot/RelationshipsCardContainer.js
@@ -1,9 +1,10 @@
 import {connect} from 'react-redux'
-import RelationshipsCard from 'snapshots/RelationshipsCard'
+import RelationshipsCard from 'common/RelationshipsCard'
 import {getPeopleSelector} from 'selectors/screening/relationshipsSelectors'
 
 const mapStateToProps = (state, _ownProps) => ({
   areRelationshipsEmpty: getPeopleSelector(state).isEmpty(),
+  isScreening: false,
 })
 
 export default connect(mapStateToProps)(RelationshipsCard)
