+++ app/javascript/screenings/RelationshipsCardContainer.jsx
@@ -1,9 +1,10 @@
 import {connect} from 'react-redux'
-import RelationshipsCard from 'screenings/RelationshipsCard'
+import RelationshipsCard from 'common/RelationshipsCard'
 import {getPeopleSelector} from 'selectors/screening/relationshipsSelectors'
 
 const mapStateToProps = (state, _ownProps) => ({
   areRelationshipsEmpty: getPeopleSelector(state).isEmpty(),
+  isScreening: true,
 })
 
 export default connect(mapStateToProps)(RelationshipsCard)
